% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{doughnut_checker}
\alias{doughnut_checker}
\title{Plot all the buffered polygons and a single object}
\usage{
doughnut_checker(
  doughnut_out,
  nbuffers,
  dist_seq,
  nobjects,
  object_n = 1,
  bg_layer = NULL
)
}
\arguments{
\item{doughnut_out}{Takes the output from \code{doughnut_builder}. An \code{sf} data frame.}

\item{nbuffers}{The number of buffers implemented.}

\item{dist_seq}{Numeric vector. Distances (in map units) for buffering each measurement point.
Each value represents the outer radius of a doughnut. Must be unique and positive.}

\item{nobjects}{The number of measurement points that were originally buffered.}

\item{object_n}{The measurement point that you want to plot separately. A single or vector of integers.}

\item{bg_layer}{Background layer to plot along with the objects. Currently only accepts \code{terra::spatRaster}.}
}
\value{
Plots showing all buffered objects and the different buffers for the object chosen in \code{object_n}.
Returns list containing ggplot objects
}
\description{
Plot all the buffered polygons and a single object
}
