% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{get_pts_buffer}
\alias{get_pts_buffer}
\title{Get number of points within varying distances}
\usage{
get_pts_buffer(x, y, poly, mindist, maxdist, incdist)
}
\arguments{
\item{x}{The measurement points}

\item{y}{The points to be summed}

\item{poly}{The polygon to be kept within}

\item{mindist}{The minimum distance}

\item{maxdist}{The maximum distance}

\item{incdist}{The distance increment}
}
\value{
list containing number of points and buffer area for every distance
increment
}
\description{
Get number of points within varying distances
}
