% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laggar_PACKAGE.R
\name{laggar-package}
\alias{laggar-package}
\alias{laggar}
\title{The 'laggar' package.}
\description{
A package for assembling data for zone-of-influence models in
mgcv

This package creates and checks data objects for fitting zone-of-influence
models, also known as scalar-on-function or lag effects models within mgcv.
The model fitting process is not itself included within this package, but
demonstrated within the introductory vignette.
}
