% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{lg_assembly}
\alias{lg_assembly}
\title{Assembling}
\usage{
lg_assembly(
  response,
  covariate,
  index_seq = NULL,
  minindex = NULL,
  maxindex = NULL,
  incindex = NULL
)
}
\arguments{
\item{response}{Vector of response values}

\item{covariate}{Matrix of covariate values, assembled by lg_ summary functions}

\item{index_seq}{Sequence of index values used to index the covariate matrix}

\item{minindex}{Minimum index, ignored if \code{index_seq} specified}

\item{maxindex}{Maximum index, ignored if \code{index_seq} specified}

\item{incindex}{index increment, ignored if \code{index_seq} specified}
}
\value{
list
}
\description{
Assembling
}
