% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lg_plot}
\alias{lg_plot}
\alias{lg_plot.sf}
\alias{lg_plot.data.frame}
\alias{lg_plot.list}
\title{Visualise scalar-on-function regression data}
\usage{
lg_plot(...)

\method{lg_plot}{sf}(data, ...)

\method{lg_plot}{data.frame}(
  data,
  response = "response",
  covariate = "covariate",
  index = "index",
  widths = c(0.1, 0.5, 0.5),
  plot_index = FALSE,
  ...
)

\method{lg_plot}{list}(
  data,
  response = "response",
  covariate = "covariate",
  index = "index",
  widths = c(0.1, 0.5, 0.5),
  plot_index = FALSE,
  ...
)
}
\arguments{
\item{...}{Other arguments passed to \code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}}

\item{data}{Either a data.frame, sf object, or a list}

\item{response}{string giving the name of the column with the response in it}

\item{covariate}{string giving the name of the matrix-column with covariate values in it}

\item{index}{string giving the name of the matrix-column with index values in it}

\item{widths}{the widths of the different plots}

\item{plot_index}{Whether to plot the index plot, by default \code{FALSE}}
}
\value{
a \code{patchwork} combination of \code{ggplot2} objects
}
\description{
Plots the response, covariate matrix and index matrix for a given SoFR term
}
\section{Methods (by class)}{
\itemize{
\item \code{lg_plot(sf)}: sf interface

\item \code{lg_plot(data.frame)}: data.frame interface

\item \code{lg_plot(list)}: list interface

}}
