% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{lg_points}
\alias{lg_points}
\title{Get number of points within varying distances}
\usage{
lg_points(
  x,
  y,
  dist_seq = NULL,
  poly = NULL,
  mindist = NULL,
  maxdist = NULL,
  incdist = NULL
)
}
\arguments{
\item{x}{The measurement points}

\item{y}{The points to be summed}

\item{dist_seq}{The sequence of distance values}

\item{poly}{Optional polygon to be kept within}

\item{mindist}{The minimum distance, ignored if dist_seq is specified}

\item{maxdist}{The maximum distance, ignored if dist_seq is specified}

\item{incdist}{The distance increment, ignored if dist_seq is specified}
}
\value{
list containing matrices of the number of points
(\code{num_points}), buffer area (\code{buffer_area}) and average number of
points per unit area  (\code{points_parea}) for every distance increment.
All matrices will have number of rows equal to the number of measurement
points unless \code{poly = NULL} in which case \code{buffer_area} and
\code{points_parea} will be returned as \code{NA}.
}
\description{
Get number of points within varying distances
}
