% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sofr_matrices.R
\name{plot_sofr_matrices}
\alias{plot_sofr_matrices}
\title{Visualize scalar-on-function regression data}
\usage{
plot_sofr_matrices(data, response, covariate, index)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{response}{string giving the name of the column with the response in it}

\item{covariate}{string giving the name of the matrix-column with covariate values in it}

\item{index}{string giving the name of the matrix-column with index values in it}
}
\value{
a \code{ggplot2} plot
}
\description{
Plots the response, covariate matrix and index matrix for a given
SoFR term in a model.
}
\author{
David L Miller
}
