% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{sum_within_dist}
\alias{sum_within_dist}
\title{Sum number of points within distance and polygon}
\usage{
sum_within_dist(x, y, poly, dist)
}
\arguments{
\item{x}{The measurement points}

\item{y}{The points to be summed}

\item{poly}{The polygon to be kept within}

\item{dist}{The distance}
}
\value{
list containing sum of points and area in buffer zone
}
\description{
Sum number of points within distance and polygon
}
